Split new-recipe file into recipe/*
