+++ runtime/recipe/base.js
@@ -1,0 +1,75 @@
+// Copyright (c) 2017 Google Inc. All rights reserved.
+// This code may only be used under the BSD style license found at
+// http://polymer.github.io/LICENSE.txt
+// Code distributed by Google as part of this project is also
+// subject to an additional IP rights grant found at
+// http://polymer.github.io/PATENTS.txt
+var assert = require('assert');
+
+function compareNulls(o1, o2) {
+  if (o1 == o2) return 0;
+  if (o1 == null) return -1;
+  return 1;
+}
+function compareStrings(s1, s2) {
+  if (s1 == null || s2 == null) return compareNulls(s1, s2);
+  return s1.localeCompare(s2);
+}
+function compareNumbers(n1, n2) {
+  if (n1 == null || n2 == null) return compareNulls(n1, n2);
+  return n1 - n2;
+}
+function compareBools(b1, b2) {
+  if (b1 == null || b2 == null) return compareNulls(b1, b2);
+  return b1 - b2;
+}
+function compareArrays(a1, a2, compare) {
+  assert(a1 != null);
+  assert(a2 != null);
+  if (a1.length != a2.length) return compareNumbers(a1.length, a2.length);
+  for (let i = 0; i < a1.length; i++) {
+    let result;
+    if ((result = compare(a1[i], a2[i])) != 0) return result;
+  }
+  return 0;
+}
+function compareObjects(o1, o2, compare) {
+  let keys = Object.keys(o1);
+  let result;
+  if ((result = compareNumbers(keys.length, Object.keys(o2).length)) != 0) return result;
+  for (let key of keys) {
+    if ((result = compare(o1[key], o2[key])) != 0) return result;
+  }
+  return 0;
+}
+function compareComparables(o1, o2) {
+  if (o1 == null || o2 == null) return compareNulls(o1, o2);
+  return o1._compareTo(o2);
+}
+
+class Node {
+  constructor(recipe) {
+    assert(recipe);
+    this._recipe = recipe;
+  }
+}
+
+class Edge {
+  constructor(recipe) {
+    assert(recipe);
+    this._recipe = recipe;
+  }
+}
+
+class Connection extends Edge {
+}
+
+exports.Node = Node;
+exports.Connection = Connection;
+exports.compareNulls = compareNulls;
+exports.compareStrings = compareStrings;
+exports.compareNumbers = compareNumbers;
+exports.compareBools = compareBools;
+exports.compareArrays = compareArrays;
+exports.compareObjects = compareObjects;
+exports.compareComparables = compareComparables;
